+++ testng-test-kit/build.gradle.kts
@@ -1,3 +1,9 @@
 plugins {
     id("testng.java-library")
 }
+
+dependencies {
+    api(projects.testngCoreApi)
+    implementation(projects.testngAsserts)
+    implementation(projects.testngCore)
+}
